no need to run readme example, just typecheck

+++ test/Churro/Test/Examples.hs
@@ -27,12 +27,12 @@ import Control.Concurrent (Chan)
 -- ** Tests
 
 -- | Example from readme:
--- >>> readme
-
+-- 
+-- No need to run it, just typecheck it.
 readme :: IO ()
 readme = do
-    runWaitChan             $ sourceList [1..10] >>> processDebug "after source" >>> delay 1 {- seconds -} >>> arr succ >>> sinkPrint
-    (wait =<<)  $ run @Chan $ sourceIO (\cb -> cb 1 >> print "Doing whatever!" >> cb 5) >>> filterC (> 3) >>> sinkIO print
+    runWaitChan             $ sourceList [1::Int ..10] >>> processDebug "after source" >>> delay 1 {- seconds -} >>> arr succ >>> sinkPrint
+    (wait =<<)  $ run @Chan $ sourceIO (\cb -> cb (1::Int) >> print "Doing whatever!" >> cb 5) >>> filterC (> 3) >>> sinkIO print
 
 -- | Checks that the IO nature of the churros doesn't duplicate operations.
 --   Actions within a pipeline should only occur once no matter how the
